[refactor] move S3 to services
